Tolerate few different frames in long similar sequence
Because of frame positioning in the timeline, it's possible that most of the frames will match and have some which will be misaligned. Over time, this can attribute to few seconds of discrepancy between the intro scenes. This algorithm "patches" the gaps between few frames to produce a continuous line.